- modify core kill script to only eliminate db instances when explicitly told to do so (-kill-db) - modify core/go.sh only bounce db instances when doing a build (skip build with -skip-build) - fix a command aggregation that bourne shell doesn't like, and include additional args for psql needed when running db in separate container

+++ docker/core/go.sh
@@ -23,11 +23,13 @@ do
   docker rm $i
 done
 
-for i in `docker ps | grep -v ^CONTAINER | grep transitclock-db- | cut -d ' ' -f1`
-do
-  docker stop $i
-  docker rm $i
-done
+if [ "$PERFORM_BUILD" == "1" ]; then
+  for i in `docker ps | grep -v ^CONTAINER | grep transitclock-db- | cut -d ' ' -f1`
+  do
+    docker stop $i
+    docker rm $i
+  done
+fi
 
 IFS=$SAVED_IFS
 
@@ -54,19 +56,17 @@ do
 
   if [ "$PERFORM_BUILD" == "1" ]; then
     docker build --no-cache -t transitclock-core-$AGENCYID .
+    docker run --name transitclock-db-$AGENCYID --rm -p $MAPPED_PORT:5432 -e POSTGRES_PASSWORD=$PGPASSWORD -d postgres:9.6.3
+
+    echo MAPPED_PORT: $MAPPED_PORT
+    DB_HOSTNAME=`docker inspect transitclock-db-$AGENCYID | grep -i ipaddress | tail -1 | cut -d '"' -f4`
+    echo DB_HOSTNAME: $DB_HOSTNAME
+
+    echo sleeping for a few...
+    sleep 5
   else
     echo skipping build...
   fi
-
-  docker run --name transitclock-db-$AGENCYID --rm -p $MAPPED_PORT:5432 -e POSTGRES_PASSWORD=$PGPASSWORD -d postgres:9.6.3
-
-  echo MAPPED_PORT: $MAPPED_PORT
-  DB_HOSTNAME=`docker inspect transitclock-db-$AGENCYID | grep -i ipaddress | tail -1 | cut -d '"' -f4`
-  # DB_HOSTNAME="0.0.0.0"
-  echo DB_HOSTNAME: $DB_HOSTNAME
-
-  echo sleeping for a few...
-  sleep 5
 
   docker run --name transitclock-core-instance-$AGENCYID --rm -e PGPASSWORD=$PGPASSWORD  -v ~/logs:/usr/local/transitclock/logs/ -v ~/ehcache:/usr/local/transitclock/cache/ transitclock-core-$AGENCYID start-core.sh $AGENCYID $GTFS_URL $GTFSRTVEHICLEPOSITIONS $PERFORM_BUILD $1 $DB_HOSTNAME 5432 $PRIMARY_AGENCY_HOST 5432 $PRIMARY_AGENCY_ID
